Convert Groovy script to Kotlin
I ran into several issues with the Groovy script. The one that stopped
me was it loaded multiple classes via several class path loaders, likely
due to the static members.

I think the original script is supposed to run in an embedded QuPath
environment but we have the opposite task: to run outside QuPath after a
DeepCell job finishes. So, I don't think we even need Groovy for
anything other than consistency … which might be a big deal, we'll see.

Meanwhile I rewrote it in Kotlin as a "Groovy-like" language (in
particular inferred types). I did a pretty literal rewrite for now, we
should refactor at some point.

Kotlin compiles the normal way, it doesn't get interpreted or have a
special runtime the way Groovy does.