Add image filter to cmdline + refactor
The filter previously known as "regionSet" is now --image-filter on the
command line. This lets the user specify which images to select, based
on containing the filter string. (Maybe we should add a regex mode in
the future… like --image-filter-re)

Also, cleans up the main class by moving the invocation argument classes
to a helper file + removing intermediate variables.