Wrap image server in full-image prefetch
The add-intensity-measurement implementation fetches the image tile
containing the region of interest … for every ROI. This means we're
reading the image over and over again. Impact: ~12.5 hours to measure
intensity for a ~140M pixel image. (‼️)

For our images so far, we can actually just load the whole image into
memory. So, wrap the image server in an adapter that prefetches the
entire image. Then, when reading the region, just fetch that region out
of the prefetched image.

This won't work when the image is too large to fit in memory. At that
point, we'll probably want to prefetch in slices, and process all ROIs
within each slice. No matter how we approach it, we don't want to fetch
tiny portions of the image repeatedly.

Fixes #7