Match nuclei by 100% inclusion in cell, or best match
We previously matched nuclei based on being fully contained in the cell.
We actually used the strict "contains" which meant the nucleus couldn't
share a border with the cell. We relaxed that to "covers" which means
they can share boundary pixels.

However we still found some nuclei weren't matching their cells because
the nucleus boundaries extended by 1-2 pixels past the cell. So now we
build a list of candidate matches based on nucleus and cell overlapping
at all, which we then sort by the percentage of the nucleus that's
contained in the cell.

The highest percentage matches are taken in order, until we have no more
matches at which point we stop.