Search for matches using quadtree spatial index
Previously, we were finding nucleus <-> cell matches with an O(n^2)
algorithm: for each nucleus, test each cell. This was very slow, taking
about 30min for a 140M pixel image with ~160k nuclei & ~160k cells. We
knew this was wasteful because we were testing cells that were nowhere
near the nucleus.

Now, we index cells with a quadtree spatial index:
https://en.wikipedia.org/wiki/Quadtree

Then, to find potential cell matches, we use the quadtree to query for
cells within the region. The quadtree might return too many matches but
won't return too few (aka it might have false positives but not false
negatives).

Otherwise the algorithm remains the same:
- find all "perfect" matches (nucleus 100% within cell)
- find all "best" matches (based on % of nucleus in cell)

This now takes ~2min on the same 140M px image.